@@ -38,8 +38,8 @@
     "mobx": "^6.2.0",
     "prettier": "^2.2.1",
     "qs": "^6.10.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-alpha-46a0f050a-20210828",
+    "react-dom": "^18.0.0-alpha-46a0f050a-20210828",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.3",
     "recoil": "^0.4.0",
